+++ src/index.js
@@ -16,7 +16,7 @@ function initialize( config, authProvider, fount ) {
 	config = config || {};
 	_.defaults( config, {
 		getUserString: function( user ) {
-			return user.name || user.username || user.id || JSON.stringify( user );
+			return user && ( user.name || user.username || user.id || JSON.stringify( user ) );
 		}
 	} );
 	authProvider = authProvider || config.authProvider;