+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.5.x
 
+### 0.6.1
+
+ * Add ability to define multiple URL patterns per action
+ * Add support for differentiated authorization (support versioned authorization in hyped)
+
 ### 0.6.0
 
 * Removed Express.js warning when not using sessions